==> code/03_executing_model.ipynb <==
# Executing fuzzy SIRDS model to data from the selected municipalities

import pandas as pd
from scipy.optimize import differential_evolution
from datetime import datetime

from model.sirds_model import get_bounds_and_arguments, sirds_objective_function

## Reading data

df = pd.read_csv('../data/df_ts_epidemic_episodes.csv', index_col=0)
df.DATA = pd.to_datetime(df.DATA)

max_date_to_fit = df.DATA.max()
max_date_to_fit

cumulative_days_in_first_outbreak_to_max_bound_I0 = 56

## Execution SIRDS model

df = df.sort_values(by=['MUNICIPIO', 'DATA'])

df.columns

df[df.MUNICIPIO == 'São Paulo'][['DATA','NOVOS_CASOS_SRAG']]

df['TAXA_CASOS_NOVOS_MEDIA_MOVEL_7_DIAS'] = df['CASOS_NOVOS_MEDIA_MOVEL_7_DIAS'] / df['POPULACAO_2022'] * 100000

try:
    df_initial_results = pd.read_csv('data/execution_results.csv')
except:
    df_initial_results = pd.DataFrame({'municipality_id':[]})

df_results = df_initial_results.copy()

DAYS_TO_RECOVERY = 8

DIFFERENTIAL_EVOLUTION_POP_SIZE_FACTOR = 5
NUMBER_ESTIMATON_PER_COUNTRY = 20

for estimation in range(NUMBER_ESTIMATON_PER_COUNTRY):        
    for municipality_id in df.CODIGO_MUNICIPIO_6.unique():       
        df_municipality = df[(df.CODIGO_MUNICIPIO_6 == municipality_id) & (df.DATA <= max_date_to_fit)].reset_index(drop=True)
        municipality = df_municipality.MUNICIPIO.unique()[0]
        print('municipality: ', municipality_id, ' ', municipality)
        
        # Dates about begin time series and first case        
        date_first_case = min(df_municipality[(df_municipality.NOVOS_CASOS_SRAG > 0)].iloc[0].DATA, df_municipality[(df_municipality.casosNovos > 0)].iloc[0].DATA)       

        # Period of analysis
        period_in_days = (max_date_to_fit - date_first_case).days + 1
                           
        estimations_performed = len(df_results[(df_results.municipality_id == municipality_id)])
    
        if (estimation == estimations_performed):
            print('estimation: '+str(estimation))
            
            rt_column = 'NUMERO_REPRODUCAO_EFETIVO_SRAG_MEDIA'
            # rt_column = 'reproduction_rate'
            
            bounds, args = get_bounds_and_arguments(df_municipality, 'DATA', 'TAXA_OBITOS_NOVOS_MEDIA_MOVEL_7_DIAS', rt_column, 'TAXA_CASOS_NOVOS_MEDIA_MOVEL_7_DIAS', 'ONSET_NUMERO_REPRODUCAO_EFETIVO_MEDIA', DAYS_TO_RECOVERY, date_first_case, max_date_to_fit, df_municipality.POPULACAO_2022.iloc[0], period_in_days, cumulative_days_in_first_outbreak_to_max_bound_I0)
            
            print(bounds)
            
            # Record the start time
            start_time = datetime.now()
            print(start_time)

            result = differential_evolution(sirds_objective_function, bounds, args=args, popsize=DIFFERENTIAL_EVOLUTION_POP_SIZE_FACTOR, maxiter=10000, workers=11, updating='deferred')
    
            # Record the end time
            end_time = datetime.now()
    
            # Calculate the duration (in seconds) for the optimization
            duration = (end_time - start_time).total_seconds()
            print(duration)
            
            list_breakpoints_in_slow_transition = args[4]
            quantity_outbreaks = args[5]
            quantity_outbreak_adjustments = args[6]
    
            # Create a dictionary to store results
            estimation_result = {
                'municipality_id': municipality_id,
                'municipality': municipality,
                'estimation': estimation,  # To differentiate between multiple estimations
                'result_fun': result.fun,
                'result_nfev': result.nfev,
                'result_nit': result.nit,
                'result_success': result.success,
                'start_time': start_time.strftime('%Y-%m-%d %H:%M:%S'),  # Format start time as a string
                'end_time': end_time.strftime('%Y-%m-%d %H:%M:%S'),  # Format end time as a string
                'duration_seconds': duration,  # Duration in seconds
                'pop_size': DIFFERENTIAL_EVOLUTION_POP_SIZE_FACTOR,
                'period_in_days': period_in_days,
                'days_to_recovery': DAYS_TO_RECOVERY,
                'cumulative_days_in_first_outbreak_to_max_bound_I0': cumulative_days_in_first_outbreak_to_max_bound_I0,
                'date_first_case': date_first_case,
                'list_breakpoints_in_slow_transition': list_breakpoints_in_slow_transition,
                'x_initial_infected_population': result.x[0],
                'x_days_between_infections_0': result.x[1]
            }
            
            quantity_epidemic_periods_with_slow_transition = len(list_breakpoints_in_slow_transition) + 1
            
            for p in range(quantity_epidemic_periods_with_slow_transition):
                estimation_result['x_case_fatality_probability_'+str(p)] = result.x[2 + p]
                estimation_result['x_loss_immunity_in_days_'+str(p)] = result.x[2 + quantity_epidemic_periods_with_slow_transition + p]                                                 
            quantity_breakpoints = (quantity_outbreaks - 1) + (quantity_outbreak_adjustments)
            begin_breakpoint_parameters = 2 + 2*quantity_epidemic_periods_with_slow_transition
            for b in range(quantity_breakpoints):
                estimation_result['x_days_between_infections_'+str(b+1)] = result.x[begin_breakpoint_parameters + b]
                estimation_result['x_breakpoint_'+str(b+1)] = result.x[begin_breakpoint_parameters + quantity_breakpoints + b]
                estimation_result['x_transition_days_between_epidemic_periods_'+str(b+1)] = result.x[begin_breakpoint_parameters + 2*quantity_breakpoints + b]
    
            print(estimation_result)
    
            # Append the estimation result to the list            
            df_results = pd.concat([df_results, pd.DataFrame.from_records([estimation_result])])
    
            df_results.to_csv('../data/execution_results.csv', index=False)           

bounds[0]

df_municipality[['date', 'rate_new_deaths_moving_average', 'rate_new_cases_moving_average', 'estimated_onset_by_deaths', 'NUMERO_REPRODUCAO_EFETIVO_ATRASADO_MEDIA']]

min(df_municipality[(df_municipality.rate_new_cases_moving_average > 0)].iloc[0].date, df_municipality[(df_municipality.estimated_onset_by_deaths > 0)].iloc[0].date)

df_municipality[(df_municipality.rate_new_cases_moving_average > 0)].iloc[0].date

df_municipality[(df_municipality.estimated_onset_by_deaths > 0)].iloc[0].date

outbreaks = df.groupby('ONSET_NUMERO_REPRODUCAO_EFETIVO_MEDIA').agg({'date': ['min', 'max']})
outbreaks

outbreaks.iloc[0][('date','min')]

rt_in_outbreak = df_municipality[
    (df_municipality['date'] >= outbreaks.iloc[0][('date', 'min')]) &
    (df_municipality['date'] <= outbreaks.iloc[0][('date', 'max')])][rt_column].values
rt_in_outbreak

df_results
