==> code/05_cross_correlation_preparing_time_series.ipynb <==
# Preparing time series for cross-correlation

import  pandas as pd

## Merging time series

df_covid_mobility = pd.read_csv('../data/df_ts_covid_mobility.csv', index_col=0)
df_covid_mobility.DATA = pd.to_datetime(df_covid_mobility.DATA, infer_datetime_format=True)

df_r0_mean = pd.read_csv('../data/df_r0_mean.csv')
df_r0_mean.date = pd.to_datetime(df_r0_mean.date)
df_r0_mean = df_r0_mean.rename(columns={'municipality_id': 'CODIGO_MUNICIPIO_6', 'date': 'DATA'})

df_municipal = pd.read_csv('../data/df_municpal.csv')

df_municipal_codes = df_municipal[['CODIGO_MUNICIPIO_6', 'CODIGO_MUNICIPIO_7']].drop_duplicates()

df_r0_mean = df_r0_mean.merge(df_municipal_codes, how='left', on='CODIGO_MUNICIPIO_6')

df_r0_mean = df_r0_mean.merge(df_covid_mobility[['CODIGO_MUNICIPIO_7', 'DATA', 'PERCENTUAL_VARIACAO_PERMANENCIA_RESIDENCIAS', 'PERCENTUAL_VARIACAO_PERMANENCIA_RESIDENCIAS_MEDIA_7_DIAS', 'NUMERO_REPRODUCAO_EFETIVO_SRAG_MEDIA', 'ONSET_NUMERO_REPRODUCAO_EFETIVO_MEDIA']], how='left', on=['CODIGO_MUNICIPIO_7', 'DATA'])

## Saving data

df_r0_mean.to_csv('../data/df_base_analysis.csv', index=False)
